gateaway/requirements.txt
pika == 1.3.2
fastapi == 0.115.12
uvicorn == 0.34.2
python-dotenv == 1.1.0